requirements.txt
rich==13.5.2
pyfiglet==0.8.post1
colorama==0.4.6
requests==2.31.0
python-dotenv==1.0.0 